Move encode/decode to separate files
